==> part2 data mining/1.autoencoder-dataset/4.missing_value_imputation_for_all_the_sample_in_dataset.ipynb <==
from google.colab import drive
drive.mount('/gdrive', force_remount=True)

import pandas as pd
import numpy as np
from sklearn.preprocessing import MinMaxScaler
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader, TensorDataset

#import data
dir = "/gdrive/My Drive/Colab Notebooks"
df = pd.read_csv('/gdrive/My Drive/Colab Notebooks/cutoff_dataset_prepare.txt', delimiter='\t')
print(df.head())
print(df.info())

#normalization
scaler = MinMaxScaler()
normalized_data = scaler.fit_transform(df)
normalized_dataframe = pd.DataFrame(normalized_data, columns=df.columns)
print(normalized_dataframe.head())

#mask out 20 percentages of entries from the complete dataset
val_dataframe = normalized_dataframe.copy()
# Calculate the average of each column
#column_means = val_dataframe.mean()

# Create a new data frame to mark modifications
mask = pd.DataFrame(1, index=val_dataframe.index, columns=val_dataframe.columns)

# Operate on each column
for column in val_dataframe.columns:
    # Determine the number of rows to replace
    num_rows = val_dataframe.shape[0]
    num_to_replace = int(num_rows * 0.20)  # 20% 的行

    # Randomly select rows
    rows_to_replace = np.random.choice(num_rows, num_to_replace, replace=False)

    # Replace the value of the selected row with 0
    val_dataframe.loc[rows_to_replace, column] = 0

    # Set the value of the corresponding position in the mask data frame to 0
    mask.loc[rows_to_replace, column] = 0

#mask = mask.astype(float)
print(val_dataframe.head())
print(mask.head())

#set auto_enconder
class HadamardAutoencoder(nn.Module):
    def __init__(self):
        super(HadamardAutoencoder, self).__init__()
        # Encoder part
        self.encoder = nn.Sequential(
            nn.Linear(in_features=428, out_features=128),
            nn.ReLU(),
            nn.Linear(128, 64)
        )
        # Decoder part
        self.decoder = nn.Sequential(
            nn.Linear(64, 128),
            nn.ReLU(),
            nn.Linear(128, 428),
            nn.Sigmoid()
        )

    def forward(self, x):
        x = self.encoder(x)
        x = self.decoder(x)
        return x


# use gpu for training if available
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print(device)

# instantiate the model
model = HadamardAutoencoder()
model.to(device) # move the model to gpu if available
print(model)
optimizer = optim.Adam(model.parameters(), lr=0.0025)
criterion = nn.L1Loss()

#Self-Supervision set
replicate_size = 20
green_points_proportion = 0.4

val_dataframe = pd.concat([val_dataframe] * replicate_size, ignore_index=True)
mask = pd.concat([mask] * replicate_size, ignore_index=True)
print(val_dataframe.info())

train_dataframe = val_dataframe.copy()
# Calculate the average of each column
column_means = train_dataframe.mean()

# Operate on each column
for column in train_dataframe.columns:
    # Find the row in this column where mask is 1
    mask_rows = mask[mask[column] == 1].index

    # Calculate the number of rows that need to be replaced (40%)
    num_to_replace = int(len(mask_rows) * green_points_proportion)

    # Randomly select some of these rows
    rows_to_replace = np.random.choice(mask_rows, num_to_replace, replace=False)

    # Replace the value of the selected row with the average value of the column
    train_dataframe.loc[rows_to_replace, column] = column_means[column]

print(train_dataframe.head())

#prepare data
val_dataframe = val_dataframe.astype(float)
train_dataframe = train_dataframe.astype(float)
mask = mask.astype(float)
validation = torch.tensor(val_dataframe.values)
train = torch.tensor(train_dataframe.values)
mask_value = torch.tensor(mask.values)
dataset = TensorDataset(train, validation, mask_value)
data_loader = DataLoader(dataset, batch_size=32, shuffle=True)
print(train_dataframe.head())
print(val_dataframe.head())

#train the model
epochs = 100
for epoch in range(epochs):
    for train_data, val_data, mask_data in data_loader:
        # Move data to the specified device (GPU or CPU)
        train_data = train_data.to(device).float()
        val_data = val_data.to(device).float()
        mask_data = mask_data.to(device).float()

        optimizer.zero_grad()
        output = model(train_data)
        loss = criterion(output * mask_data, val_data * mask_data)  # Apply Hadamard product
        loss.backward()
        optimizer.step()

    if epoch % 1 == 0:
        print(f'Epoch {epoch}, Loss: {loss.item()}')

train = train.float()
final = model(train)
loss = criterion(final * mask_value, validation * mask_value)
print(f'Loss: {loss.item()}')

loss=criterion(train * mask_value, validation * mask_value)
print(f'Loss: {loss.item()}')

#deal with normalized_dataframe
normalized_dataframe = pd.concat([normalized_dataframe] * replicate_size, ignore_index=True)
norm = torch.tensor(normalized_dataframe.values)
loss=criterion(validation , norm )
print(f'Loss: {loss.item()}')
#This is the loss of average imputation

validation = validation.float()
impute = model(validation)
impute_norm = torch.where(mask_value == 1, validation, impute)
loss=criterion(impute_norm , norm )
print(f'Loss: {loss.item()}')
#This is the loss of the imputation we use

#read original data
origin = pd.read_csv('/gdrive/My Drive/Colab Notebooks/cutoff_dataset_final.txt', delimiter='\t')
mask_na = (~origin.isna()).astype(int)
origin.fillna(df.mean(), inplace=True)
print(origin.info())
print(origin.head())
print(mask_na.info())
print(mask_na.head())

#normalization
scaler = MinMaxScaler()
normalized_origin = scaler.fit_transform(origin)
normalized_origin = pd.DataFrame(normalized_origin, columns=origin.columns)
print(normalized_origin.head())


origin_tensor = torch.tensor(normalized_origin.values)
cover = torch.tensor(mask_na.values)

origin_tensor = origin_tensor.float()
origin_impute = model(origin_tensor)
new = torch.where(cover == 1, origin_tensor, origin_impute)

new_detached = new.detach()
new_np = new_detached.numpy()
new_dataframe = pd.DataFrame(new_np, columns=origin.columns)
print(new_dataframe.head())
print(normalized_origin.head())

new_dataframe.to_csv('/gdrive/My Drive/Colab Notebooks/dataset_output.txt', sep='\t', index=False)
